feat: adiciona tela de splash screen
